+++ WAMVoiceCapture/TDLibBridge.h
@@ -1,0 +1,6 @@
+#ifndef WAM_VOICE_CAPTURE_TDLIB_BRIDGE_H
+#define WAM_VOICE_CAPTURE_TDLIB_BRIDGE_H
+
+#include <td/telegram/td_json_client.h>
+
+#endif